Update the develop branch

+++ ServicesApplicationFrontend/Screens/Register.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { StyleSheet, View } from 'react-native';
-import { Text, TextInput, Button, Switch, HelperText, Menu, Divider } from 'react-native-paper';
+import { Alert, StyleSheet, View } from 'react-native';
+import { Text, TextInput, Button, Switch, HelperText, Menu, Divider, ActivityIndicator } from 'react-native-paper';
 
 import axios from 'axios';
 import AsyncStorage from '@react-native-async-storage/async-storage';
@@ -9,45 +9,50 @@ export default function Register({ navigation }) {
 
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [confirmPassword, setConfirmPassword] = useState('')
   const [accountCreated, setAccountCreated] = useState(false);
+  const [loading, setLoading] = useState(false);
 
   const handleRegister = async () => {
     //Call /api/register
     //Email verify 
     //Login user by placing user tokens in AsyncStorage
-
-
-    setAccountCreated(!accountCreated);
+    setLoading(true);
     console.log(accountCreated);
     if (accountCreated) {
 
     } else {
 
-      axios.post('http://192.168.100.140:3000/api/register', { email, password })
-        .then(response => {
-          //tell user to approve account via email
-          //try to logIn
+      if (password != confirmPassword) {
+        Alert.alert('Passwords need to match');
+      } else {
 
-          console.log(response);
+        axios.post('http://192.168.96.112:3000/api/register', { email, password })
 
 
-        }).catch(err => {
-          //somehow alert the user there's an error
-          console.error(err);
-        })
+          .then(response => {
+            //tell user to approve account via email
+            //try to logIn
+            console.log(response);
+            setAccountCreated(!accountCreated);
 
+          }).catch(err => {
+            //somehow alert the user there's an error
+            console.error(err);
+          })
+      }
     }
-
-
   }
 
-  const hangeleRegisterNext = async () => {
+  const handleRegisterNext = async () => {
+
     //try to logIn
-    axios.post('http://192.168.100.140:3000/api/login', { email, password })
+    axios.post('http://192.168.100.99:3000/api/login', { email, password })
+
       .then(response => {
         //go to build profile
         //store details in async storage
-        AsyncStorage.setItem('UserDetails',JSON.stringify(response));
+        AsyncStorage.setItem('UserDetails', JSON.stringify(response));
         navigation.push('BuildProfileScreen')
       }).catch(err => {
         //somehow alert the user there's an error
@@ -62,24 +67,43 @@ export default function Register({ navigation }) {
           Create An account
         </Text>
       </View>
-      <TextInput
-        style={{ ...styles.input, backgroundColor: "white" }}
-        value={email}
-        label='email'
-        onChangeText={(text) => setEmail(text)}
-      />
+      {loading ? (
+        <>
+          <ActivityIndicator animating={true} />
+        </>
+      ) : (
+        <>
+          <TextInput
+            style={{ ...styles.input, backgroundColor: "white" }}
+            value={email}
+            label='email'
+            onChangeText={(text) => setEmail(text)}
+          />
 
-      <TextInput
-        style={{ ...styles.input, backgroundColor: "white" }}
-        value={password}
-        label='password'
-        onChangeText={(text) => setPassword(text)}
-        secureTextEntry={true}
-      />
+          <TextInput
+            style={{ ...styles.input, backgroundColor: "white" }}
+            value={password}
+            label='password'
+            onChangeText={(text) => setPassword(text)}
+            secureTextEntry={true}
+          />
 
-      <Button mode='contained' style={styles.input} onPress={() => handleRegister()} > Send Verification </Button>
-      <Button style={styles.input} onPress={() => navigation.push('LoginScreen')}> Login </Button>
-      {accountCreated ? <Button style={styles.input} mode='elevated' onPress={() => doLogin()}> Next </Button> : <></>}
+          <TextInput
+            style={{ ...styles.input, backgroundColor: "white" }}
+            value={confirmPassword}
+            label='confirm password'
+            onChangeText={(text) => setConfirmPassword(text)}
+            secureTextEntry={true}
+          />
+
+          <Text style={styles.Information}> The Password should be at least 8 characters long </Text>
+
+          <Button mode='contained' style={styles.input} onPress={() => handleRegister()} > Send Email Verification </Button>
+          <Button style={styles.input} onPress={() => navigation.push('LoginScreen')}> Login </Button>
+        </>
+      )}
+
+      {accountCreated ? <Button style={styles.input} mode='elevated' onPress={() => handleRegisterNext()}> Next </Button> : <></>}
     </View>
   )
 }
@@ -93,6 +117,10 @@ const styles = StyleSheet.create({
     marginVertical: 20,
     justifyContent: "space-between",
   },
-  textContainer: { alignContent: 'center', alignItems: 'center' }
+  textContainer: { alignContent: 'center', alignItems: 'center' },
+  Information: {
+    color: 'purple',
+    fontSize: 15
+  }
 
 });
